Nuxt 3 as `nuxt`, Nuxt 2 as `nuxt2`
