+++ package.json
@@ -18,26 +18,26 @@
     "next.js",
     "angular",
     "nuxt",
-    "nuxt3",
+    "nuxt2",
     "ssr",
     "vite"
   ],
   "sideEffects": false,
   "scripts": {
     "build": "rimraf dist && tsc",
-    "test": "npm run test:angular && npm run test:next && npm run test:nuxt && npm run test:nuxt3 && npm run test:custom",
+    "test": "npm run test:angular && npm run test:next && npm run test:nuxt2 && npm run test:nuxt && npm run test:custom",
     "e2e:build-next": "rimraf e2e/next-test && cd e2e && npx create-next-app next-test --use-npm --ts && cd next-test && cp ../../tools/firebase.json . && cp ../../tools/.firebaserc .",
     "e2e:build-vite": "rimraf e2e/vite-test && cd e2e && npx create-vite vite-test --template vue && cd vite-test && npm i && cp ../../tools/firebase.json . && cp ../../tools/.firebaserc .",
-    "e2e:build-nuxt": "rimraf e2e/nuxt-test && cd e2e && npx create-nuxt-app nuxt-test --answers '$(cat ../tools/create-nuxt-answers.json)' && cd nuxt-test && cp ../../tools/firebase.json . && cp ../../tools/.firebaserc .",
-    "e2e:build-nuxt3": "rimraf e2e/nuxt3-test && cd e2e && npx nuxi init nuxt3-test && cd nuxt3-test && npm i && cp ../../tools/firebase.json . && cp ../../tools/.firebaserc .",
+    "e2e:build-nuxt2": "rimraf e2e/nuxt2-test && cd e2e && npx create-nuxt-app nuxt2-test --answers '$(cat ../tools/create-nuxt-answers.json)' && cd nuxt2-test && cp ../../tools/firebase.json . && cp ../../tools/.firebaserc .",
+    "e2e:build-nuxt": "rimraf e2e/nuxt-test && cd e2e && npx nuxi init nuxt-test && cd nuxt-test && npm i && cp ../../tools/firebase.json . && cp ../../tools/.firebaserc .",
     "e2e:build-angular": "rimraf e2e/angular-test && cd e2e && npm i -g @angular/cli && ng new angular-test --defaults --skip-git && cd angular-test && cp ../../tools/firebase.json . && cp ../../tools/.firebaserc .",
     "e2e:build-custom": "rimraf e2e/custom/lib && rimraf e2e/custom/.firebase && cd e2e/custom && npm i && cp ../../tools/firebase.json . && cp ../../tools/.firebaserc .",
     "clean:build": "rimraf dist",
     "test:angular": "npm run e2e:build-angular && node --loader ts-node/esm ./tools/test-angular.ts",
     "test:vite": "npm run e2e:build-vite && node --loader ts-node/esm ./tools/test-vite.ts",
     "test:next": "npm run e2e:build-next && node --loader ts-node/esm ./tools/test-next.ts",
+    "test:nuxt2": "npm run e2e:build-nuxt2 && node --loader ts-node/esm ./tools/test-nuxt2.ts",
     "test:nuxt": "npm run e2e:build-nuxt && node --loader ts-node/esm ./tools/test-nuxt.ts",
-    "test:nuxt3": "npm run e2e:build-nuxt3 && node --loader ts-node/esm ./tools/test-nuxt3.ts",
     "test:custom": "npm run e2e:build-custom && node --loader ts-node/esm ./tools/test-custom.ts"
   },
   "exports": {
@@ -92,8 +92,8 @@
     "firebase-tools": "^11.14.4",
     "inquirer": "^8.2.0",
     "next": "13.4.7",
-    "nuxt": "^2.15.8",
-    "nuxt3": "npm:nuxt@^3.4.3",
+    "nuxt2": "npm:nuxt@^2.15.8",
+    "nuxt": "^3.4.3",
     "rimraf": "^3.0.2",
     "sharp": "^0.32.1",
     "ts-node": "^10.9.1",
