+++ src/nuxt2/firebase-aware.ts
