+++ tools/test-nuxt.ts
@@ -1,7 +1,7 @@
 import { exit } from 'process';
+import { existsSync, readFileSync } from 'fs';
 import { readdir, access } from 'fs/promises';
 import { join } from 'path';
-import { existsSync, readFileSync } from 'fs';
 import { execSync } from 'child_process';
 
 const site = 'nextjs-demo-73e34';
@@ -13,7 +13,6 @@ const run = async () => {
     if (await access(join(cwd, '.firebase', site)).then(() => false, () => true)) throw `.firebase/${site} does not exist`;
     if (await access(join(cwd, '.firebase', site, 'hosting')).then(() => false, () => true)) throw `.firebase/${site}/hosting does not exist`;
     if (!(await readdir(join(cwd, '.firebase', site, 'hosting'))).length) throw `no files in .firebase/${site}/hosting`;
-    if (await access(join(cwd, '.firebase', site, 'functions')).then(() => true, () => false)) throw `.firebase/${site}/functions should not exist`;
 }
 
 run().then(
