+++ tools/test-nuxt2.ts
@@ -1,11 +1,11 @@
 import { exit } from 'process';
-import { existsSync, readFileSync } from 'fs';
 import { readdir, access } from 'fs/promises';
 import { join } from 'path';
+import { existsSync, readFileSync } from 'fs';
 import { execSync } from 'child_process';
 
 const site = 'nextjs-demo-73e34';
-const cwd = join('e2e', 'nuxt3-test');
+const cwd = join('e2e', 'nuxt2-test');
 
 const run = async () => {
     execSync('firebase emulators:exec "exit 0"', { cwd });
@@ -13,6 +13,7 @@ const run = async () => {
     if (await access(join(cwd, '.firebase', site)).then(() => false, () => true)) throw `.firebase/${site} does not exist`;
     if (await access(join(cwd, '.firebase', site, 'hosting')).then(() => false, () => true)) throw `.firebase/${site}/hosting does not exist`;
     if (!(await readdir(join(cwd, '.firebase', site, 'hosting'))).length) throw `no files in .firebase/${site}/hosting`;
+    if (await access(join(cwd, '.firebase', site, 'functions')).then(() => true, () => false)) throw `.firebase/${site}/functions should not exist`;
 }
 
 run().then(
